Moved logic to other file
